Add next-intl localization
