+++ middleware.ts
@@ -1,0 +1,8 @@
+import createMiddleware from 'next-intl/middleware'
+import {routing} from './src/i18n/routing'
+
+export default createMiddleware(routing)
+
+export const config = {
+  matcher: ['/((?!_next|.*\\..*).*)'],
+}
