+++ src/app/[locale]/promotion/page.tsx
