+++ src/components/Footer.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import Image from 'next/image'
-import Link from 'next/link'
+import {Link} from '@/i18n/navigation'
 import { usePathname } from 'next/navigation'
 import {
   FaPhoneAlt,
