+++ src/components/PopularServices.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react'
 import { motion } from 'framer-motion'
-import Link from 'next/link'
+import {Link} from '@/i18n/navigation'
 import Image from 'next/image'
 
 const services = [
