+++ src/components/PrimaryButton.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Link from 'next/link'
+import {Link} from '@/i18n/navigation'
 import { ButtonHTMLAttributes, ReactNode } from 'react'
 
 interface PrimaryButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
