+++ src/components/navbar/Navbar.tsx
@@ -9,6 +9,7 @@ import {
   useLayoutEffect,
 } from 'react'
 import { usePathname } from 'next/navigation'
+import {useTranslations} from 'next-intl'
 import MegaMenu from './MegaMenu'
 import LanguageSwitcher from './LanguageSwitcher'
 import HamburgerButton from './HamburgerButton'
@@ -19,6 +20,7 @@ import CustomLink from '@/components/CustomLink'
 import { FaFire } from 'react-icons/fa'
 
 export default function Navbar() {
+  const t = useTranslations('navbar')
   const [isHoveringMenu, setIsHoveringMenu] = useState(false)
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false)
   const hideTimer = useRef<NodeJS.Timeout | null>(null)
@@ -109,13 +111,13 @@ export default function Navbar() {
           <CustomLink
             href="/promotion"
             section="Navbar"
-            item="โปรโมชั่น"
+            item={t('promotion')}
             className="flex items-center gap-1 font-medium text-black hover:text-[#A70909] transition relative
               after:content-[''] after:absolute after:bottom-[-24px] after:left-0 after:w-full after:h-[7px]
               after:bg-[#A70909] after:rounded-t-full after:scale-x-0 hover:after:scale-x-120 after:origin-middle after:transition-transform"
           >
             <FaFire className="text-[#A70909] animate-bounce" />
-            โปรโมชั่น
+            {t('promotion')}
           </CustomLink>
 
           {/* 🧾 บริการ (เมก้าเมนู) */}
@@ -124,12 +126,12 @@ export default function Navbar() {
             onMouseLeave={handleMouseLeave}
             className="relative"
           >
-            <span className={`relative transition  
+            <span className={`relative transition
               after:content-[''] after:absolute after:bottom-[-24px] after:left-0 after:w-full after:h-[7px]
               after:bg-[#A70909] after:rounded-t-full
               after:transition-transform after:origin-middle
               ${isHoveringMenu ? 'text-[#A70909] after:scale-x-120' : 'text-black after:scale-x-0'}`}>
-              บริการ
+              {t('services')}
             </span>
 
             {isHoveringMenu && (
@@ -149,26 +151,26 @@ export default function Navbar() {
           <CustomLink
             href="/under-construction"
             section="Navbar"
-            item="ดาวน์โหลดเอกสาร"
+            item={t('download')}
             className="relative text-black hover:text-[#A70909] transition
               after:content-[''] after:absolute after:bottom-[-24px] after:left-0 after:w-full after:h-[7px]
               after:bg-[#A70909] after:rounded-t-full
               after:scale-x-0 hover:after:scale-x-120 after:origin-middle after:transition-transform"
           >
-            ดาวน์โหลดเอกสาร
+            {t('download')}
           </CustomLink>
 
           {/* ☎️ ติดต่อเรา */}
           <CustomLink
             href="/under-construction"
             section="Navbar"
-            item="ติดต่อเรา"
+            item={t('contact')}
             className="relative text-black hover:text-[#A70909] transition
               after:content-[''] after:absolute after:bottom-[-24px] after:left-0 after:w-full after:h-[7px]
               after:bg-[#A70909] after:rounded-t-full
               after:scale-x-0 hover:after:scale-x-120 after:origin-middle after:transition-transform"
           >
-            ติดต่อเรา
+            {t('contact')}
           </CustomLink>
         </nav>
