+++ src/i18n/navigation.ts
@@ -1,0 +1,4 @@
+import {createNavigation} from 'next-intl/navigation'
+import {routing} from './routing'
+
+export const {Link, usePathname, useRouter, redirect} = createNavigation(routing)
