+++ src/i18n/routing.ts
@@ -1,0 +1,9 @@
+import {defineRouting} from 'next-intl/routing'
+
+export const locales = ['th', 'en'] as const
+export const defaultLocale = 'th'
+
+export const routing = defineRouting({
+  locales,
+  defaultLocale,
+})
