+++ src/messages/en.json
@@ -1,0 +1,12 @@
+{
+  "hero": {
+    "title": "Full-Service Accounting & Business Partner",
+    "subtitle": "Fast, easy, all-in-one business solutions"
+  },
+  "navbar": {
+    "promotion": "Promotion",
+    "services": "Services",
+    "download": "Downloads",
+    "contact": "Contact Us"
+  }
+}
