+++ src/messages/th.json
@@ -1,0 +1,12 @@
+{
+  "hero": {
+    "title": "สำนักงานบัญชีและบริหารธุรกิจครบวงจร",
+    "subtitle": "บริการไว ใช้งานง่าย ครบจบทุกเรื่องธุรกิจ"
+  },
+  "navbar": {
+    "promotion": "โปรโมชั่น",
+    "services": "บริการ",
+    "download": "ดาวน์โหลดเอกสาร",
+    "contact": "ติดต่อเรา"
+  }
+}
